requirements.txt

fastapi>=0.111.0
uvicorn==0.24.0
pymongo==4.6.0
motor==3.3.2
python-dotenv==1.0.0
pydantic>=2.6.4
pydantic-settings>=2.2.1
python-multipart==0.0.7
pytest==7.4.0
pytest-asyncio==0.22.0
slowapi==0.1.6
pyjwt
httpx==0.24.1
razorpay==2.0.0
